+++ src/components/item/ItemImageViewer.js
@@ -1,15 +1,32 @@
 /* @flow */
 
 import React, {Component} from 'react';
-import {View, StyleSheet, TouchableHighlight} from 'react-native';
+import {
+  View,
+  StyleSheet,
+  TouchableOpacity,
+  StatusBar,
+  Animated,
+  Text,
+  SafeAreaView,
+  TouchableWithoutFeedback,
+} from 'react-native';
 
 //library
-import Icon from 'react-native-vector-icons/FontAwesome';
+import Ionicons from 'react-native-vector-icons/Ionicons';
 import {Actions} from 'react-native-router-flux';
 import ImageViewer from 'react-native-image-zoom-viewer';
 import EventTracker from '../../helper/EventTracker';
-import {saveImage} from '../../helper/image';
-import ActionSheet from 'react-native-actionsheet';
+// import ActionSheet from 'react-native-actionsheet';
+import appConfig from 'app-config';
+import {HIT_SLOP} from 'app-packages/tickid-chat/constants';
+import RightButtonNavBar from '../RightButtonNavBar';
+import {RIGHT_BUTTON_TYPE} from '../RightButtonNavBar/constants';
+
+const HEADER_BUTTON_TYPE = {
+  BACK_BUTTON: 'back_button',
+  DOWNLOAD_IMAGE_BUTTON: 'download_image_button',
+};
 
 class ItemImageViewer extends Component {
   static defaultProps = {
@@ -19,13 +36,15 @@ class ItemImageViewer extends Component {
   constructor(props) {
     super(props);
 
-    this.imageIndex = this.props.index;
+    this.isHeaderVisible = true;
+    this.opacity = new Animated.Value(1);
     this.OPTIONS_LIST = [
       props.t('common:saveImageLabel'),
       props.t('common:cancel'),
     ];
 
     this.refActionSheet = React.createRef();
+    this.refButtonDownloadImage = React.createRef();
   }
 
   eventTracker = new EventTracker();
@@ -38,22 +57,91 @@ class ItemImageViewer extends Component {
     this.eventTracker.clearTracking();
   }
 
-  handleCloseImageView = () => Actions.pop();
+  toggleHeaderVisibility = () => {
+    this.isHeaderVisible = !this.isHeaderVisible;
+    Animated.timing(this.opacity, {
+      toValue: this.isHeaderVisible ? 1 : 0,
+      duration: 400,
+      useNativeDriver: true,
+    }).start();
+  };
+  
+  handlePressHeaderBtn = (type) => {
+    if (!this.isHeaderVisible) {
+      this.handlePressImage();
+      return;
+    }
+    switch (type) {
+      case HEADER_BUTTON_TYPE.BACK_BUTTON:
+        this.handleBack();
+        break;
+      case HEADER_BUTTON_TYPE.DOWNLOAD_IMAGE_BUTTON:
+        this.handleDownloadImage();
+      default:
+        break;
+    }
+  };
 
-  handleOnSwipeDownImage = () => Actions.pop();
+  handleDownloadImage = () => {
+    if (this.refButtonDownloadImage.current) {
+      this.refButtonDownloadImage.current.handlePressDownloadImage();
+    }
+  };
 
-  handleOnLongPressImage = () => {
+  handleBack = () => {
+    Actions.pop();
+  };
+
+  handleSwipeDownImage = () => Actions.pop();
+
+  handleLongPressImage = () => {
     if (this.refActionSheet.current) this.refActionSheet.current.show();
   };
 
-  onChangeImageIndex = (index) => {
-    this.imageIndex = index;
-  };
+  renderIndicator = () => null;
 
-  handleOptionPress = (index) => {
-    if (index !== this.OPTIONS_LIST.length - 1) {
-      saveImage(this.props.images[this.imageIndex].url);
-    }
+  renderHeader = (currentIndex) => {
+    return (
+      <TouchableWithoutFeedback onPress={this.toggleHeaderVisibility}>
+        <Animated.View
+          style={[styles.headerContainer, {opacity: this.opacity}]}>
+          <SafeAreaView style={styles.headerContentContainter}>
+            <View style={styles.headerContent}>
+              <TouchableOpacity
+                hitSlop={HIT_SLOP}
+                style={styles.headerLeftButton}
+                onPress={() =>
+                  this.handlePressHeaderBtn(HEADER_BUTTON_TYPE.BACK_BUTTON)
+                }>
+                <Ionicons size={26} name="ios-chevron-back" color="#fff" />
+              </TouchableOpacity>
+              <View style={styles.headerMiddleContainer}>
+                <Text style={styles.headerMiddleTitle}>
+                  {currentIndex + 1}/{this.props.images.length}
+                </Text>
+              </View>
+
+              <TouchableOpacity
+                onPress={() =>
+                  this.handlePressHeaderBtn(
+                    HEADER_BUTTON_TYPE.DOWNLOAD_IMAGE_BUTTON,
+                  )
+                }>
+                <View pointerEvents="none">
+                  <RightButtonNavBar
+                    ref={this.refButtonDownloadImage}
+                    touchableOpacity
+                    type={RIGHT_BUTTON_TYPE.DOWNLOAD_IMAGE}
+                    containerStyle={styles.headerRightBtnContainer}
+                    imageUrl={this.props.images[currentIndex].url}
+                  />
+                </View>
+              </TouchableOpacity>
+            </View>
+          </SafeAreaView>
+        </Animated.View>
+      </TouchableWithoutFeedback>
+    );
   };
 
   render() {
@@ -61,30 +149,28 @@ class ItemImageViewer extends Component {
 
     return (
       <View style={styles.container}>
+        <StatusBar hidden />
         <ImageViewer
+          style={styles.imageViewerContainer}
+          index={this.props.index}
+          imageUrls={images}
+          saveToLocalByLongPress={false}
+          renderHeader={this.renderHeader}
+          renderIndicator={this.renderIndicator}
           enableSwipeDown={true}
           swipeDownThreshold={100}
-          onSwipeDown={this.handleOnSwipeDownImage}
-          saveToLocalByLongPress={false}
-          onLongPress={this.handleOnLongPressImage}
-          imageUrls={images}
-          index={this.props.index}
-          onChange={this.onChangeImageIndex}
+          onSwipeDown={this.handleSwipeDownImage}
+          onClick={this.toggleHeaderVisibility}
+          onLongPress={this.handleLongPressImage}
         />
 
-        <TouchableHighlight
-          onPress={this.handleCloseImageView}
-          style={styles.btnCloseImageView}>
-          <Icon name="times-circle" size={32} color="#ffffff" />
-        </TouchableHighlight>
-
-        <ActionSheet
+        {/* <ActionSheet
           ref={this.refActionSheet}
           options={this.OPTIONS_LIST}
           cancelButtonIndex={1}
           destructiveButtonIndex={1}
           onPress={this.handleOptionPress}
-        />
+        /> */}
       </View>
     );
   }
@@ -92,6 +178,11 @@ class ItemImageViewer extends Component {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
+  },
+  imageViewerContainer: {
+    position: 'absolute',
+    width: appConfig.device.width,
+    height: appConfig.device.height,
   },
   btnCloseImageView: {
     position: 'absolute',
@@ -101,6 +192,38 @@ const styles = StyleSheet.create({
     height: 50,
     justifyContent: 'center',
     alignItems: 'center',
+  },
+  headerContainer: {
+    flexDirection: 'row',
+    position: 'absolute',
+    zIndex: 999,
+    backgroundColor: 'rgba(0,0,0,0.65)',
+    alignItems: 'center',
+  },
+  headerContentContainter: {
+    flex: 1,
+  },
+  headerContent: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  headerLeftButton: {
+    padding: 20,
+  },
+  headerMiddleContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  headerMiddleTitle: {
+    color: '#fff',
+    fontSize: 16,
+  },
+  headerRightBtnContainer: {
+    paddingRight: 20,
+    paddingTop: 20,
+    paddingLeft: 20,
+    paddingBottom: 20,
   },
 });