+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - PersonalProfile and UserChat from skv.
 ### Changed
+- Update image button download mechanism and back button in item ImageViewer by [kienluu612@gmail.com].
 - Only use has_attr for checking product's attributes.
 - Handle onclick of <a> tag and window.open event of react-native-auto-height-webview.
 ### Fixed